chore: 首页预览修复 + 侧边栏分组 + MeiliSearch PageIndex 清理
- 首页: 资源卡片点击预览（FilePreviewComponent），仅显示可预览资源
- 后端: 公开浏览 API 过滤草稿状态资源
- 侧边栏: 重组成 5 个分组栏目 + 分割线 CSS
- 导航: 新增 5 个本地化 key（资源管理 / AI 管理 / 专业和课程管理 / 实训和就业管理 / 资讯和双高信息管理）
- MeiliSearch: 移除 PageIndex 卡片和相关代码
- 修复: 多项预存构建错误（TypeScript + C#）
- 新增: ChapterResource 实体/服务/迁移、StudentProfile 组件、AI 工具类

+++ src/KnowledgeHub.Application/Learning/StudentExerciseRecordAppService.cs
@@ -488,7 +488,7 @@ public class StudentExerciseRecordAppService : KnowledgeHubAppService, IStudentE
             TotalStudents = allLearnerIds,
             ActiveStudents = activeStudentIds,
             TotalExercises = allExercises.Count,
-            TotalLearningMinutes = Math.Round((decimal)totalTimeMinutes, 1),
+            TotalLearningMinutes = (double)Math.Round((decimal)totalTimeMinutes, 1),
             AverageCompletionRate = allExercises.Count > 0 && allLearnerIds > 0
                 ? Math.Round((decimal)completedRecords.Count / (allExercises.Count * allLearnerIds) * 100, 1)
                 : 0,
